[mugen] deserialize state controller persistent values. fix deserialization of maps. use an integer for the random index value instead of a char

+++ src/mugen/random.h
@@ -34,7 +34,7 @@ protected:
     void init();
 
     uint64_t state[16];
-    uint8_t index;
+    uint32_t index;
 
     static PaintownUtil::ReferenceCount<Random> current;
 };
